Further removed common code out of the publish task.
The code in question is mostly to deal with general "submissions".
It will be reused in the flight task.
